Feture fe/test thang

+++ MarvicSolution/MarvicSolution.BackendApi/Controllers/ProjectController.cs
@@ -1,4 +1,5 @@
 ﻿using MarvicSolution.BackendApi.Constants;
+using MarvicSolution.DATA.Enums;
 using MarvicSolution.Services.Issue_Request.Dtos.ViewModels;
 using MarvicSolution.Services.Project_Request.Project_Resquest;
 using MarvicSolution.Services.Project_Request.Project_Resquest.Dtos;
@@ -24,7 +25,7 @@ namespace MarvicSolution.BackendApi.Controllers
 
         // /api/Project/GetAlls
         [HttpGet]
-        [Route("/api/Project/GetAlls")] // remember to check this route
+        [Route("/api/Project/GetAlls")]
         public async Task<IActionResult> GetAlls()
         {
             if (!ModelState.IsValid)
@@ -96,17 +97,22 @@ namespace MarvicSolution.BackendApi.Controllers
         /// <param name="rq">Request from client</param>
         /// <returns></returns>
         [HttpPost]
-        [Route("/api/Project/Create")]// remember to check this route
+        [Route("/api/Project/Create")]
         public async Task<IActionResult> Create([FromBody] Project_CreateRequest rq)
         {
             // Check model state
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-            // Create a project
-            var IdProj = await _projectService.Create(UserLogin.Id, rq);
-            if (IdProj.Equals(Guid.Empty))
-                return BadRequest();
-            return Ok("Create project success");
+            if (UserLogin.Role.Equals(2) || UserLogin.Role.Equals(1))
+            {
+                // Create a project
+                var IdProj = await _projectService.Create(UserLogin.Id, rq);
+                if (IdProj.Equals(Guid.Empty))
+                    return BadRequest();
+                return Ok("Create project success");
+            }
+            else
+                return Content("You do not have permission to perform this function");
         }
 
         // api/Project/AddMember?IdProject=xxx-xxx-xx
@@ -114,12 +120,17 @@ namespace MarvicSolution.BackendApi.Controllers
         [Route("/api/Project/AddMember")]
         public IActionResult AddMember([FromBody] AddMember_Request rq)
         {
-            if (!ModelState.IsValid)
-                return BadRequest(ModelState);
-            var idProject = _projectService.AddMembers(rq.IdProject, rq.UserNames, UserLogin.Id);
-            if (idProject.Equals(Guid.Empty))
-                return BadRequest($"Cannot get projects = {rq.IdProject}");
-            return Ok(idProject);
+            if (UserLogin.Role.Equals(2) || UserLogin.Role.Equals(1))
+            {
+                if (!ModelState.IsValid)
+                    return BadRequest(ModelState);
+                var idProject = _projectService.AddMembers(rq.IdProject, rq.UserNames, UserLogin.Id);
+                if (idProject.Equals(Guid.Empty))
+                    return BadRequest($"Cannot get projects = {rq.IdProject}");
+                return Ok(idProject);
+            }
+            else
+                return Content("You do not have permission to perform this function");
         }
 
         // api/Project/RemoveMember
@@ -129,13 +140,18 @@ namespace MarvicSolution.BackendApi.Controllers
         {
             try
             {
-                if (!ModelState.IsValid)
-                    return BadRequest(ModelState);
+                if (UserLogin.Role.Equals(2) || UserLogin.Role.Equals(1))
+                {
+                    if (!ModelState.IsValid)
+                        return BadRequest(ModelState);
 
-                var result = _projectService.Remove_Member_From_Project(rq.IdProject, rq.IdUser, UserLogin.Id);
-                if (result.Equals(Guid.Empty))
-                    return BadRequest($"Cannot remove idUser = {rq.IdUser} from IdProject = {rq.IdProject}");
-                return Ok(result);
+                    var result = _projectService.Remove_Member_From_Project(rq.IdProject, rq.IdUser, UserLogin.Id);
+                    if (result.Equals(Guid.Empty))
+                        return BadRequest($"Cannot remove idUser = {rq.IdUser} from IdProject = {rq.IdProject}");
+                    return Ok(result);
+                }
+                else
+                    return Content("You do not have permission to perform this function");
             }
             catch (Exception e)
             {
@@ -144,18 +160,23 @@ namespace MarvicSolution.BackendApi.Controllers
 
         }
         [HttpPut]
-        [Route("/api/Project/Update")]// remember to check this route
+        [Route("/api/Project/Update")]
         public async Task<IActionResult> Update([FromBody] Project_UpdateRequest rq)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-            var result = await _projectService.Update(UserLogin.Id, rq);
-            if (result.Equals(Guid.Empty))
-                return BadRequest();
-            return Ok("Update project success");
+            if (UserLogin.Role.Equals(2) || UserLogin.Role.Equals(1))
+            {
+                var result = await _projectService.Update(UserLogin.Id, rq);
+                if (result.Equals(Guid.Empty))
+                    return BadRequest();
+                return Ok("Update project success");
+            }
+            else
+                return Content("You do not have permission to perform this function");
         }
         [HttpPatch]
-        [Route("/api/Project/UpdateStarredProject")]// remember to check this route
+        [Route("/api/Project/UpdateStarredProject")]
         public async Task<IActionResult> UpdateStarredProject([FromBody] UpdateStarredProject_Request rq)
         {
             if (!ModelState.IsValid)
@@ -166,15 +187,18 @@ namespace MarvicSolution.BackendApi.Controllers
             return Ok(result);
         }
         [HttpPatch]
-        [Route("/api/Project/DisableMember")]// remember to check this route
-        public IActionResult DisableMember([FromBody] DisableMember_Request rq)
+        [Route("/api/Project/ChangeStatusMember")]
+        public IActionResult ChangeStatusMember([FromBody] ChangeStatusMember_Request rq)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-            var result = _projectService.DisableMember(rq);
-            if (!result)
-                return BadRequest();
-            return Ok(result);
+            if (UserLogin.Role.Equals(2) || UserLogin.Role.Equals(1))
+            {
+                var result = _projectService.ChangeStatusMember(rq);
+                return Ok(result);
+            }
+            else
+                return Content("You do not have permission to perform this function");
         }
 
         [HttpDelete("{proj_Id}")]
@@ -182,10 +206,15 @@ namespace MarvicSolution.BackendApi.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-            var affectedResutl = await _projectService.Delete(proj_Id, UserLogin.Id);
-            if (affectedResutl.Equals(Guid.Empty))
-                return BadRequest();
-            return Ok("Delete project success");
+            if (UserLogin.Role.Equals(2) || UserLogin.Role.Equals(1))
+            {
+                var affectedResutl = await _projectService.Delete(proj_Id, UserLogin.Id);
+                if (affectedResutl.Equals(Guid.Empty))
+                    return BadRequest();
+                return Ok("Delete project success");
+            }
+            else
+                return Content("You do not have permission to perform this function");
         }
 
         // api/Project/GetStarredProject
@@ -200,5 +229,25 @@ namespace MarvicSolution.BackendApi.Controllers
             //    return BadRequest($"Cannot get list username by IdProject = {UserLogin.Id}");
             return Ok(project);
         }
+
+        // api/Project/SetUserRoleByIdProject
+        [HttpGet]
+        [Route("/api/Project/SetUserRoleByIdProject")]
+        public IActionResult SetUserRoleByIdProject(Guid idProject)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var rq = new SetUserRoleByIdProject_Request(UserLogin.Id, idProject);
+            var resultVM = _projectService.SetUserRoleByIdProject(rq);
+
+            if (resultVM == null)
+                return BadRequest();
+            // assign role for user login
+            UserLogin.Role = resultVM.Value;
+
+            return Ok(resultVM);
+        }
+
     }
 }
